+++ packages/client/base/src/types/index.ts
@@ -2,6 +2,7 @@ export * from "./events";
 export * from "./hosted";
 export * from "./embed";
 export * from "./system";
+export * from "./payment-method-management";
 export * from "./CryptoCurrency";
 
 export const BaseUrls = {