+++ packages/client/ui/react-native/src/components/index.ts
@@ -1,1 +1,3 @@
 export * from "./embed";
+export * from "./signers";
+export * from "./wallets";