+++ packages/client/ui/react-ui/src/components/index.ts
@@ -4,5 +4,7 @@ export * from "./CrossmintNFTDetail";
 export * from "./embed";
 export * from "./hosted";
 export * from "./signers";
+export * from "./wallets";
+export * from "./card-management";
 
 export { EmbeddedAuthForm } from "./auth/EmbeddedAuthForm";