+++ packages/wallets/src/wallets/wallet-factory.ts
@@ -46,7 +46,9 @@ export class WalletFactory {
             );
         }
 
-        const existingWallet = await this.apiClient.getWallet(`me:${this.getChainType(args.chain)}:smart`);
+        const locator = this.getWalletLocator<C>(args);
+
+        const existingWallet = await this.apiClient.getWallet(locator);
 
         if (existingWallet != null && !("error" in existingWallet)) {
             return this.createWalletInstance(existingWallet, args);
@@ -120,6 +122,7 @@ export class WalletFactory {
                 ...(delegatedSigners != null ? { delegatedSigners } : {}),
             },
             owner: args.owner ?? undefined,
+            alias: args.alias ?? undefined,
         } as CreateWalletParams);
 
         if ("error" in walletResponse) {
@@ -157,6 +160,7 @@ export class WalletFactory {
                     args.options?.shadowSignerStorage
                 ),
                 options: args.options,
+                alias: args.alias,
             },
             this.apiClient
         );
@@ -264,6 +268,10 @@ export class WalletFactory {
             return delegatedSigner;
         }
         throw new WalletCreationError(`${signerLocator} signer does not match the wallet's signer type`);
+    }
+
+    private getWalletLocator<C extends Chain>(args: WalletArgsFor<C>): string {
+        return `me:${this.getChainType(args.chain)}:smart` + (args.alias != null ? `:alias:${args.alias}` : "");
     }
 
     private async createPasskeySigner<C extends Chain>(