+++ packages/client/react-base/CHANGELOG.md
@@ -1,5 +1,34 @@
 # @crossmint/client-sdk-react-base
 
+## 0.7.2
+
+### Patch Changes
+
+- Updated dependencies [2cae54d]
+  - @crossmint/client-sdk-base@1.7.3
+  - @crossmint/client-sdk-auth@1.2.36
+  - @crossmint/common-sdk-auth@1.0.58
+  - @crossmint/wallets-sdk@0.18.1
+
+## 0.7.1
+
+### Patch Changes
+
+- Updated dependencies [8fc9ed9]
+- Updated dependencies [28e6fd3]
+  - @crossmint/wallets-sdk@0.18.0
+
+## 0.7.0
+
+### Minor Changes
+
+- 4ab93e0: Add wallet aliasing
+
+### Patch Changes
+
+- Updated dependencies [4ab93e0]
+  - @crossmint/wallets-sdk@0.17.0
+
 ## 0.6.9
 
 ### Patch Changes